@markhpc 2022-07-22
Some generic 1 OSD sanity benchmarks:

RBD 4KB IOPS (higher is better)
test | main | wip-faster-rocksdb
-- | -- | --
randread | 114597 | 113057
randwrite | 75530 | **94072**
read | 28192 | 29888
write | 298033 | 322859

RBD 4KB Cycles/OP (lower is better)
test | main | wip-faster-rocksdb
-- | -- | --
randread | 221096 | 224686
randwrite | 585403 | **560758**
read | 189125 | 186563
write | 30013 | **26119**

RBD 4KB Lat (lower is better)
test | main (avg ms) | main (99% ms) | wip-faster-rocksdb (avg ms) | wip-faster-rocksdb (99% ms)
-- | -- | -- | -- | --
randread | 4.701 | 43.385 | 4.812 | 41.353
randwrite | 8.065 | 67.158 | **6.060** | **34.275**
read | 19.359 | 163.414 | **17.319** | **90.735**
write | 1.714 | 3.113 | **1.583** | **2.695**

RGW:
test | main | wip-faster-rocksdb
-- | -- | --
puts/second | 5536 | 5515
list time (seconds) | **593** | 718
gets/second | 12743 | 12760
deletes/second | 7187 | **7832**

Overall write and delete gains as expected (though we may need more RGW instances to see the write benefit in the hsbench test).  bucket list performance is down, which may be consistent with extra overhead of reads in a bigger L0 with more data and overlapping key ranges (some potential evidence of this also with limited CPU and compression enabled on in the blog article).  It's possible that on a real cluster with multiple OSDs we don't see the effect of the list performance impact due to bucket index sharding across multiple OSDs.