@markhpc 2022-07-28
@baergj yeah, my hope is that we can perhaps back this off even farther if the other methods to compact during iteration and compact/flush on hitting a certain number of deletes between compactions.  Maybe we only fall back on that every 24 hours or something.

Generally speaking we are going to increase write amp here though.  One possible way to claw some of that back could be compression.  I don't think it's helping with write-amp much for RBD, but RGW it appeared to have a pretty massive effect.