@baergj 2022-07-28
I wonder if this is starting to approach the limit of what makes sense for index OSDs vs. data OSDs, given the vastly different workloads between the two of them?

FWIW, I had also wondered about backing TTL off to 24h+ for data OSDs, but we haven't had much reason to apply TTL to our data OSDs. IIRC rocksdb had a 30day default TTL; not sure if that's still there at HEAD.