@yaarith 2022-12-12
@ljflores @markhpc interesting, looks like the original commit (https://github.com/ceph/ceph/pull/47001/commits/b8bbf642c1bc985e5e08639f44adc0e0d7c686da) was merged in July 21, when Mark created this PR, so this might have been a rebase/fork issue. I'll check to see that there are no issues with the second commit in the telemetry code.