@yaarith 2022-12-14
The telemetry code looks ok to me.